+++ OneBlink.SDK/models/FormSubmissionEventConfigurationMappingDependentFieldValue.cs
@@ -1,0 +1,19 @@
+namespace OneBlink.SDK.Model
+{
+
+    public class FormSubmissionEventConfigurationMappingDependentFieldValue
+    {
+        public string category
+        {
+            get; set;
+        }
+        public string subCategory
+        {
+            get; set;
+        }
+        public string item
+        {
+            get; set;
+        }
+    }
+}